refactor: 共通レイアウトの整理とファイル、ディレクトリ管理の改善
