+++ docker/mysql/data/binlog.index
@@ -4,3 +4,4 @@
 ./binlog.000004
 ./binlog.000005
 ./binlog.000006
+./binlog.000007
