+++ src/app/Http/Requests/ContactRequest.php
@@ -51,20 +51,20 @@ class ContactRequest extends FormRequest
             'email.required' => 'メールアドレスを入力してください',
             'email.email' => 'メールアドレスはメール形式で入力してください',
             'email.max' => 'メールアドレスを255文字以下で入力してください',
-            'phone1.required' => '電話番号を入力してください',
-            'phone1.numeric' => '電話番号を数値で入力してください',
-            'phone1.digits_between' => '電話番号は5桁までの数字で入力してください',
-            'phone2.required' => '電話番号を入力してください',
-            'phone2.numeric' => '電話番号を数値で入力してください',
-            'phone2.digits_between' => '電話番号は5桁までの数字で入力してください',
-            'phone3.required' => '電話番号を入力してください',
-            'phone3.numeric' => '電話番号を数値で入力してください',
-            'phone3.digits_between' => '電話番号は5桁までの数字で入力してください',
+            'phone1.required' => '電話番号(1)を入力してください',
+            'phone1.numeric' => '電話番号(1)を数値で入力してください',
+            'phone1.digits_between' => '電話番号(1)は5桁までの数字で入力してください',
+            'phone2.required' => '電話番号(2)を入力してください',
+            'phone2.numeric' => '電話番号(2)を数値で入力してください',
+            'phone2.digits_between' => '電話番号(2)は5桁までの数字で入力してください',
+            'phone3.required' => '電話番号(3)を入力してください',
+            'phone3.numeric' => '電話番号(3)を数値で入力してください',
+            'phone3.digits_between' => '電話番号(3)は5桁までの数字で入力してください',
             'address.required' => '住所を入力してください',
             'address.string' => '住所を文字列で入力してください',
-            'address.name.max' => '住所を255文字以下で入力してください',
+            'address.max' => '住所を255文字以下で入力してください',
             'building.string' => '建物名を文字列で入力してください',
-            'building.name.max' => '建物名を255文字以下で入力してください',
+            'building.max' => '建物名を255文字以下で入力してください',
             'content.required' => 'お問い合わせの種類を選択してください',
             'detail.required' => 'お問い合わせ内容を入力してください',
             'detail.string' => 'お問い合わせ内容を文字列で入力してください',
