+++ src/app/Providers/FortifyServiceProvider.php
@@ -44,7 +44,6 @@ class FortifyServiceProvider extends ServiceProvider
         });
 
         // カスタムルート(登録処理)
-        // Route::post('/register', [RegisteredUserController::class, 'store'])->name('register');
         Route::middleware(['web'])->group(function () {
             Route::post('/register', [RegisteredUserController::class, 'store'])->name('register');
         });
@@ -55,38 +54,5 @@ class FortifyServiceProvider extends ServiceProvider
 
             return Limit::perMinute(10)->by($email . $request->ip());
         });
-
-        // ログアウト後のリダイレクト先を設定
-        // Fortify::logoutView(function () {
-        //     return redirect('auth.login');
-        // });
-
-        // ログイン後のリダイレクト先を設定
-        // Fortify::authenticateUsing(function (Request $request) {
-        //     $credentials = $request->only('email', 'password');
-
-        //     if (Auth::attempt($credentials)) {
-        //     return Auth::user();
-        //     }
-
-        //     return null;
-        // });
-
-        // // ログイン後のリダイレクト先を指定
-        // Fortify::redirects('login', '/admin');
-
-        // Fortify::updateUserProfileInformationUsing(UpdateUserProfileInformation::class);
-        // Fortify::updateUserPasswordsUsing(UpdateUserPassword::class);
-        // Fortify::resetUserPasswordsUsing(ResetUserPassword::class);
-
-        // RateLimiter::for('login', function (Request $request) {
-        //     $throttleKey = Str::transliterate(Str::lower($request->input(Fortify::username())).'|'.$request->ip());
-
-        //     return Limit::perMinute(5)->by($throttleKey);
-        // });
-
-        // RateLimiter::for('two-factor', function (Request $request) {
-        //     return Limit::perMinute(5)->by($request->session()->get('login.id'));
-        // });
     }
 }
