+++ src/database/seeders/DatabaseSeeder.php
@@ -13,8 +13,6 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        // \App\Models\User::factory(10)->create();
-
         $this->call(CategorySeeder::class);
 
         $this->call(ContactSeeder::class);
