--- src/public/css/confirm.css
@@ -1,85 +1,0 @@
-/* 基本設定 */
-body {
-  font-family: 'Playfair Display', serif;
-  color: #6a4e3b;
-}
-
-/* ヘッダー */
-.header {
-  padding: 15px 0;
-  border-bottom: 1px solid #eee;
-  text-align: center;
-}
-
-.header__title {
-  margin: 0;
-  font-size: 2rem;
-}
-
-/* メインコンテンツ */
-.confirm__container {
-  width: 80%;
-  max-width: 1230px;
-  margin: 0 auto;
-}
-
-.confirm__title {
-  text-align: center;
-  margin: 40px 0;
-  font-size: 1.8rem;
-}
-
-.confirm__table {
-  width: 100%;
-  border-collapse: collapse;
-  margin-bottom: 30px;
-  border: 1px solid #eee;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-}
-
-.confirm__row {
-  border-bottom: 1px solid #ddd;
-}
-
-.confirm__label {
-  background-color: #957760;
-  padding: 15px 0 15px 40px;
-  width: 30%;
-  color: #fff;
-  text-align: left;
-}
-
-.confirm__value {
-  padding: 15px 0 15px 40px;
-  width: 70%;
-  text-align: left;
-}
-
-.confirm__buttons {
-  text-align: center;
-}
-
-.confirm__button {
-  width: 100px;
-  padding: 5px 10px;
-  border: none;
-  cursor: pointer;
-  font-size: 1rem;
-  color: #fff;
-  background-color: #957760;
-  transition: background-color 0.3s;
-}
-
-.confirm__button:hover {
-  background-color: #6a4e3b;
-}
-
-.confirm__link {
-  color: inherit;
-  margin-left: 40px;
-  text-decoration: none;
-}
-
-.confirm__link:hover {
-  text-decoration: underline;
-}
