+++ src/public/css/contact/index.css
@@ -1,0 +1,196 @@
+/* ヘッダー */
+.header {
+  border-bottom: 1px solid #eee;
+}
+
+/* メインコンテンツ */
+.form {
+  width: 80%;
+  max-width: 1000px;
+  margin: 0 auto;
+}
+
+.form__title {
+  text-align: center;
+  margin: 40px 0;
+  font-size: 1.8rem;
+}
+
+.form__group {
+  margin-bottom: 30px;
+  display: flex;
+  align-items: center;
+}
+
+.form__error {
+  color: red;
+  font-size: 0.875rem;
+  margin-top: 0.25rem;
+}
+
+.form__label-box {
+  width: 35%;
+}
+
+.form__label {
+  width: 100%;
+}
+
+.form__input-box {
+  width: 65%;
+}
+
+
+.form__required {
+  color: red;
+}
+
+.form__group--email .form__input,
+.form__group--address .form__input,
+.form__group--building .form__input {
+  width: 100%;
+}
+
+.form__input,
+.form__select,
+.form__textarea {
+  padding: 10px 0;
+  border: none;
+  color: #6a4e3b;
+  background-color: #f2f2f2;
+}
+
+.form__select,
+.form__textarea {
+  padding-left: 20px;
+}
+
+.form__input:not(.form__input--phone) {
+  padding-left: 20px;
+}
+
+
+/* お名前 */
+.form__group--name .form__input-box {
+  display: flex;
+  justify-content: space-between;
+}
+
+.form__group--name .form__input {
+  width: 45%;
+}
+
+/* 性別 */
+.form__radio-label {
+  margin-right: 50px;
+}
+
+.form__radio {
+  display: none;
+}
+
+.form__radio + .form__radio-label {
+  position: relative;
+  padding-left: 30px;
+  cursor: pointer;
+  font-size: 1rem;
+  line-height: 1.5;
+}
+
+.form__radio + .form__radio-label::before {
+  content: "";
+  position: absolute;
+  left: 0;
+  top: 50%;
+  transform: translateY(-50%);
+  width: 20px;
+  height: 20px;
+  border: 2px solid #6a4e3b;
+  border-radius: 50%;
+  background-color: #fff;
+  transition: background-color 0.3s ease, border-color 0.3s ease;
+}
+
+.form__radio:checked + .form__radio-label::before {
+  background-color: #6a4e3b;
+  border-color: #6a4e3b;
+}
+
+.form__radio:checked + .form__radio-label::after {
+  content: "";
+  position: absolute;
+  left: 5px;
+  top: 50%;
+  transform: translateY(-50%);
+  width: 10px;
+  height: 10px;
+  background-color: #6a4e3b;
+  border-radius: 50%;
+}
+
+/* 電話番号 */
+.form__group--phone .form__input-box {
+  display: flex;
+  justify-content: space-between;
+}
+
+.form__input--phone {
+  width: 25%;
+  text-align: center;
+}
+
+/* お問い合わせの種類 */
+.form__group--type .form__input-box {
+  position: relative;
+  width: 35%;
+}
+
+.form__select {
+  width: 100%;
+  height: 40px;
+  padding: 0 40px 0 20px;
+  appearance: none;
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  position: relative;
+}
+
+.form__group--type .form__input-box::after {
+  content: "";
+  position: absolute;
+  right: 10px;
+  top: 50%;
+  transform: translateY(-50%);
+  border-left: 8px solid transparent;
+  border-right: 8px solid transparent;
+  border-top: 8px solid #6a4e3b;
+  pointer-events: none;
+}
+
+/* お問い合わせ内容 */
+.form__group--content {
+  align-items: start;
+}
+
+.form__group--content .form__textarea {
+  width: 100%;
+  resize: none;
+  overflow: auto;
+}
+
+/* 確認画面ボタン */
+.form__button {
+  margin: 10px auto;
+  width: 130px;
+  padding: 8px;
+  color: #fff;
+  border: none;
+  cursor: pointer;
+  font-size: 1rem;
+  background-color: #957760;
+  transition: background-color 0.3s;
+}
+
+.form__button:hover {
+  background-color: #6a4e3b;
+}
