--- src/public/css/register.css
@@ -1,102 +1,0 @@
-/* 基本設定 */
-body {
-  font-family: 'Playfair Display', serif;
-  background-color: #f5ebe0;
-  color: #6a4e3b;
-}
-
-/* ヘッダー */
-.header {
-  background-color: #fff;
-  padding: 15px 0;
-  text-align: center;
-}
-
-.header__title {
-  font-size: 2rem;
-  margin: 0;
-}
-
-.header__login-link {
-  position: absolute;
-  top: 25px;
-  right: 50px;
-  font-size: 1rem;
-  color: #8c8373;
-  background-color: #eee;
-  border: 1px solid #ddd;
-  width: 100px;
-  padding: 3px 15px;
-  text-decoration: none;
-}
-
-.header__login-link:hover {
-  text-decoration: underline;
-}
-
-/* メインコンテンツ */
-.register__title {
-  font-size: 1.8rem;
-  margin: 40px 0;
-  text-align: center;
-}
-
-.register {
-  background-color: #fff;
-  padding: 30px;
-  border-radius: 8px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  width: 600px;
-  margin:0 auto;
-}
-
-/* フォーム */
-.register__form-container {
-  width: 100%;
-}
-
-.form__group {
-  width: 70%;
-  margin: 30px auto;
-}
-
-.form__label {
-  display: block;
-  margin-bottom: 5px;
-  font-size: 1rem;
-}
-
-.form__input {
-  width: 100%;
-  padding: 12px;
-  border: none;
-  background-color: #f7f7f7;
-  font-size: 0.8rem;
-  color: #6a4e3b;
-}
-
-.form__error {
-  color: red;
-  font-size: 0.9rem;
-  margin-top: 5px;
-}
-
-.form__group--submit {
-  width: 20%;
-  margin-top: 60px;
-}
-
-.form__button {
-  width: 100%;
-  padding: 6px 12px;
-  background-color: #8c8373;
-  color: white;
-  border: none;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.form__button:hover {
-  background-color: #6f655a;
-}
