--- src/resources/views/thanks.blade.php
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Thanks - FashionablyLate</title>
-    <link href="https://fonts.googleapis.com/css2?family=Playfair+Display:wght@400;700&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="{{ asset('css/sanitize.css') }}">
-    <link rel="stylesheet" href="{{ asset('css/thanks.css') }}">
-</head>
-
-<body>
-    <div class="thanks-page__container">
-        <h1 class="thanks-page__background-text">Thank you</h1>
-        <div class="thanks-page__content">
-            <p class="thanks-page__message">お問い合わせありがとうございました</p>
-            <a class="thanks-page__button" href="/">HOME</a>
-        </div>
-    </div>
-</body>
-
-</html>
